bugfix(workflow) fix incorrect db query for status to product relation
