+++ module/workflow/src/Domain/Query/ProductStatusQueryInterface.php
@@ -10,11 +10,12 @@ declare(strict_types=1);
 namespace Ergonode\Workflow\Domain\Query;
 
 use Ergonode\SharedKernel\Domain\Aggregate\ProductId;
+use Ergonode\SharedKernel\Domain\Aggregate\StatusId;
 
-interface ProductWorkflowStatusQueryInterface
+interface ProductStatusQueryInterface
 {
     /**
-     * @return string[]
+     * @return ProductId[]
      */
-    public function getStatuses(ProductId $productId): array;
+    public function findProductIdsByStatusId(StatusId $statusId): array;
 }
